services/worker/pyproject.toml
[project]
name = "dprk-cti-worker"
version = "0.1.0"
description = "Prefect worker scaffold for DPRK CTI platform"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27.2",
  "pydantic-settings>=2.6.1",
  "prefect>=2.20.0",
  "opentelemetry-api>=1.27",
  "opentelemetry-sdk>=1.27",
  "opentelemetry-exporter-otlp-proto-grpc>=1.27",
  # Bootstrap ETL (PR #5 / PR #6)
  "openpyxl>=3.1.5",
  "pyyaml>=6.0.2",
  "pydantic>=2.9",
  "sqlalchemy[asyncio]>=2.0.36",
  "psycopg[binary]>=3.2.3",
  # aiosqlite is required at runtime so `python -m worker.bootstrap`
  # `--dry-run` without a configured database can fall back to an
  # in-memory sqlite engine. The cost is ~30KB of pure-Python wheel;
  # production runs against Postgres ignore this driver entirely.
  "aiosqlite>=0.20",
  # RSS ingest (PR #8)
  "feedparser>=6.0.11",
]

[project.scripts]
worker-main = "worker.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/worker"]

[tool.hatch.build.targets.wheel.force-include]
# Ship the shared alias dictionary inside the worker wheel so
# `python -m worker.bootstrap --workbook ...` works from an
# installed package without requiring --aliases-path. The
# repo-checkout path resolves to `data/dictionaries/aliases.yml`
# at the repo root; this entry mirrors that file into the wheel
# at `worker/bootstrap/data/aliases.yml`.
"../../data/dictionaries/aliases.yml" = "worker/bootstrap/data/aliases.yml"
"../../data/dictionaries/feeds.yml" = "worker/ingest/data/feeds.yml"
"../../data/dictionaries/taxii_collections.yml" = "worker/ingest/taxii/data/taxii_collections.yml"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-ra", "--strict-markers", "--strict-config"]

[tool.uv]
package = true
